+++ app/models/conversation.rb
@@ -1,4 +1,6 @@
 class Conversation < ApplicationRecord
+  has_secure_token :inbound_email_token
+
   belongs_to :developer
   belongs_to :business
 
@@ -29,6 +31,10 @@ class Conversation < ApplicationRecord
 
   def hiring_fee_eligible?
     developer_replied? && created_at <= 2.weeks.ago
+  end
+
+  def mark_notifications_as_read(user)
+    notifications_as_conversation.where(recipient: user).unread.mark_as_read!
   end
 
   private